Parted createResultContainer and listenForClick functions from RCC
After parting these function I also added animation trigger for tooltip and moved svg code from the function to the file root above
