+++ src/ts/RCC.ts
@@ -2,6 +2,7 @@ import RGB from "./components/rgb";
 import Hex from "./components/hex";
 import HSL from "./components/hsl";
 import generateDOMTree from "./components/generateDOMTree";
+import { createResultContainer } from "./components/createResultContainer";
 
 export default class RCC {
   private _controlTag: HTMLElement;
@@ -137,7 +138,7 @@ export default class RCC {
       }
       this._createRandomColor()
         .then((res) => {
-          this._createResultContainer(res);
+          createResultContainer(res, this._colorForm);
         })
         .catch((err) => {
           throw new Error(err);
@@ -161,53 +162,5 @@ export default class RCC {
       const errorMessage = "Something went wrong";
       reject(errorMessage);
     });
-  }
-
-  private _createResultContainer(newColor: string) {
-    if (this._colorForm.nextElementSibling)
-      this._colorForm.nextElementSibling.remove();
-
-    const justBakedContainer = new DocumentFragment();
-
-    const section = document.createElement("section");
-    section.classList.add("fresh-baked");
-    const h3 = document.createElement("h2");
-    h3.innerText = "Here, this is your new color! 😊";
-
-    const div = document.createElement("div");
-    div.classList.add("fresh-baked__block-label");
-    const colorPreviewContainer = document.createElement("div");
-    colorPreviewContainer.classList.add("fresh-baked__color-preview");
-    colorPreviewContainer.style.setProperty(
-      "--custom-background-color",
-      newColor
-    );
-    const outputElement = document.createElement("p");
-    outputElement.classList.add("fresh-baked__output-element");
-    outputElement.innerText = newColor;
-    const pasteToClipboardButton = document.createElement("button");
-    pasteToClipboardButton.innerHTML = `<svg xmlns="http://www.w3.org/2000/svg" width="36" height="36" viewBox="0 0 24 24" style="fill: rgba(0, 0, 0, 1);transform: ;msFilter:;"><path d="M19 3h-2.25a1 1 0 0 0-1-1h-7.5a1 1 0 0 0-1 1H5c-1.103 0-2 .897-2 2v15c0 1.103.897 2 2 2h14c1.103 0 2-.897 2-2V5c0-1.103-.897-2-2-2zm0 17H5V5h2v2h10V5h2v15z"></path></svg>`;
-    pasteToClipboardButton.classList.add("fresh-baked__clipboard-button");
-    // button.onclick
-    this._listenForClick(pasteToClipboardButton, newColor);
-
-    div.append(colorPreviewContainer);
-    div.append(outputElement);
-    div.append(pasteToClipboardButton);
-    section.append(h3);
-    section.append(div);
-    justBakedContainer.appendChild(section);
-    this._colorForm.after(justBakedContainer);
-
-    // creating a mini container that shows the generated value to the user with ability to copy it and a mini block to show the color live.
-  }
-
-  private _listenForClick(
-    pasteToClipboardButton: HTMLButtonElement,
-    newColor: string
-  ) {
-    pasteToClipboardButton.onclick = () => {
-      navigator.clipboard.writeText(newColor);
-    };
   }
 }
